% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_attributes}
\alias{get_attributes}
\title{Get attributes}
\usage{
get_attributes(dataset)
}
\arguments{
\item{dataset}{the slug of the dataset}
}
\description{
This function returns a list of attributes
}
\examples{
get_attributes(dataset = "insee-act-trim-anc")
}
\keyword{dataset}
