% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_attributes_keys}
\alias{get_attributes_keys}
\title{Get attributes keys}
\usage{
get_attributes_keys(dataset)
}
\arguments{
\item{dataset}{the slug of the dataset}
}
\description{
This function returns a vector of slug of attributes
}
\examples{
get_attributes_keys(dataset = "insee-act-trim-anc")
}
\keyword{dataset}

