% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider.R
\name{get_datasets}
\alias{get_datasets}
\title{Get datasets for one provider}
\usage{
get_datasets(provider)
}
\arguments{
\item{provider}{the slug of the provider}
}
\description{
This function returns a dataframe with all datasets for one provider
}
\examples{
get_datasets(provider = "insee") 
}
\keyword{providers}

