% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_dimensions_keys}
\alias{get_dimensions_keys}
\title{Get dimensions keys}
\usage{
get_dimensions_keys(dataset)
}
\arguments{
\item{dataset}{the slug of the dataset}
}
\description{
This function returns a list of dimensions
}
\examples{
get_dimensions_keys("insee-cho-an-halo")
}
\keyword{dataset}
