% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_frequencies}
\alias{get_frequencies}
\title{Get frequencies}
\usage{
get_frequencies(slug)
}
\arguments{
\item{slug}{the slug of the dataset}
}
\description{
This function returns a vector with all possible frequencies for a dataset
}
\examples{

get_frequencies(slug = "insee-act-trim-anc")

}
\keyword{dataset}
