% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_providers.R
\name{get_providers_table}
\alias{get_providers_table}
\title{Get Providers Table}
\usage{
get_providers_table()
}
\arguments{
\item{no}{parameter}
}
\description{
This function returns a data frame with data on providers
}
\examples{
get_providers_table()
}
\keyword{providers}

