% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serie.R
\name{get_serie}
\alias{get_serie}
\title{Get a serie}
\usage{
get_serie(serie)
}
\arguments{
\item{serie}{the slug of a serie}
}
\description{
This function returns the values of a serie
}
\examples{
get_serie(serie = "bis-pp-ls-q-au")
}
\keyword{serie}

