% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_series}
\alias{get_series}
\title{Get series of a dataset}
\usage{
get_series(slug)
}
\arguments{
\item{dataset}{the slug of the dataset}
}
\description{
This function returns a data frame with all series for a dataset
}
\examples{
get_series(slug = "insee-act-trim-anc")

}
\keyword{dataset}
