% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_.R
\name{select_data}
\alias{select_data}
\title{Select data}
\usage{
select_data(x)
}
\arguments{
\item{x}{a list}
}
\description{
This function selects the data part of the output
}
\examples{
library(jsonlite)
library(magrittr)
"providers" \%>\% 
 paste0(widukind_api, .) \%>\% 
 fromJSON() \%>\%
 select_data()
}
\keyword{providers}
