% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_.R
\name{select_values}
\alias{select_values}
\title{Select values}
\usage{
select_values(x)
}
\arguments{
\item{x}{a list}
}
\description{
This function extract values
}
\examples{
library(jsonlite)
library(magrittr)
"bis-pp-ls-q-au" \%>\% 
 paste0(widukind_api, "series/", .) \%>\% 
 fromJSON() \%>\% 
 select_data() \%>\% 
 select_values() 
}

