% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider.R
\name{get_datasets_keys}
\alias{get_datasets_keys}
\title{Get keys of a dataset}
\usage{
get_datasets_keys(provider)
}
\arguments{
\item{provider}{the slug of the provider}
}
\description{
This function returns a vector with all datasets keys for one provider
}
\examples{
get_datasets_keys(provider = "insee") 
}
\keyword{providers}

